Move SortOrder to types, clean a bunch of imports
